requirements.txt
asgiref==3.7.2
dj-rest-auth==5.1.0
Django==3.2.25
django-cors-headers==4.1.0
django-filter==23.5
djangorestframework==3.15.1
drf-nested-routers==0.93.4
importlib-metadata==6.7.0
install==1.3.5
jsonpickle==3.0.4
psycopg2==2.9.9
python-dotenv==0.21.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.15.0
